pyproject.toml
[project]
name = "projectpublicnotice"
version = "0.1.0"
description = "POC de extração e estruturação de editais de fomento com LLM-as-a-judge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "pdfplumber>=0.10.0",
    "pydantic>=2.5.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
    "firecrawl-py>=1.0.0",
    "pyinstaller>=6.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
]
